Rename file field-with-label
